Move util packages to sub-package w.r.t type
