+++ noty-android/app/simpleapp/src/main/java/dev/shreyaspatil/noty/simpleapp/view/notes/NotesFragment.kt
@@ -34,6 +34,10 @@ import dev.shreyaspatil.noty.simpleapp.view.base.BaseFragment
 import dev.shreyaspatil.noty.simpleapp.view.hiltNotyMainNavGraphViewModels
 import dev.shreyaspatil.noty.simpleapp.view.notes.adapter.NotesListAdapter
 import dev.shreyaspatil.noty.utils.*
+import dev.shreyaspatil.noty.utils.ext.hide
+import dev.shreyaspatil.noty.utils.ext.setDrawableLeft
+import dev.shreyaspatil.noty.utils.ext.shareWhileObserved
+import dev.shreyaspatil.noty.utils.ext.show
 import dev.shreyaspatil.noty.view.viewmodel.NotesViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
